fix(tests): apply Holly R1 nits — dedup withMigrationHome + makeMalformedPack into shared fixtures
Holly R1 findings (2 nits, applied):

1. `withMigrationHome` was duplicated in `pai-migration-issue-102.test.ts`
   and `pai-migration-issue-112.test.ts`. Promoted to shared
   `test/fixtures/pai-migration-fixtures.ts` as
   `withMigrationHome(fn, prefix?)`, wrapping `withTempHome` +
   identity-fixture write + packs-dir derivation. Both test files now
   alias-import it locally with their per-file prefix.

2. `makeMalformedPack` was copy-pasted in `pai-migration-issue-97.test.ts`,
   `pai-migration-issue-102.test.ts`, and `pai-migration-issue-112.test.ts`.
   Promoted to shared fixtures as `writeMalformedPaiPack(packsDir, packName)`.
   All three files now alias-import it.

Tests: 743/743 pass. Typecheck clean.
